Add virtual attributes for Opening and Closing DateTime
Add a custom validator to check that timeframe and DateTime values are
valid. Add validation to ensure uniqueness of reference token. Check the
presence of attachement and validity of timeframe and DateTime values at
the time of publishing of the notice. Use reference_token as the default
param for TenderNotice model. Set the DateTime values as nil if passed
values are invalid.